app: Add overlaid widget for in-app notifications
GdNotification does not really handle multiple (stacked) notifications
visible at a time, so we roll our own. It should be pretty close to
the implementation in Gd, except for the nice drop shadows.

We want to add app notifications from more places than the main
window, so it's easier to handle as a public property on the
application object.