mainWindow: Simplify updates on room changes
With PolariRoom:channel now being construct-only, we no longer have
to care for the case of a room becoming available before its channel.
TpChannel:connection is also construct-only, so we can just access it
when we have a valid room.
The same doesn't apply for TpConnection's account, but let's try to
get away with it anyway (if it goes away, we are likely on our way to
be invalidated).