chatroomManager: Select the first room from the observer
While we are joined to any rooms, there should be a currently
active one; so while we generally only move the selection when
we are actually handling a channel, make an exception if the newly
observed room is the only one. In most cases the room has been
requested by ourselves, except in the case where a contact opens
a private conversation before we join any rooms.