ui: Move border between view and input entry
The border is currently implemented as bottom-border on the view;
however when we'll adapt the selection-mode patter, the room list
will gain a toolbar at the bottom. As the toolbar is only visible in
selection mode, the border between room list and toolbar needs to
be implemented as top-border on the latter. In order to align with
the border between view and input entry, we need to use a top-border
in the latter case as well.