ui: Change leave button in toolbar to selection mode toggle
Allan suggested to use the selection mode pattern instead of a direct
leave action in the header. It is a bit odd that joining a room involves
an extra step, but leaving it (as a destructive action) does not, so let's
try this. On the other hand, leaving more than a single room at once is
very uncommon, so we might reconsider this change in the future.