appNotifications: Add 2nd notification queue for local command output
We will add support at least for some IRC commands via command line
shortly. In case a command has some outputs, most IRC clients just
dump the local output to the main chat view. This is not a very
compelling solution, as it blurs the distinction between remotely
visible content and strictly local output. So instead, we will use
a dedicated area similar to in-app notifications.