mainWindow: Implement nick completion
Add simple tab completion that cycles through matches on each
Tab press while displaying possible completions in a popup
(except in case of a single match). The Escape key may be used
to cancel the operation.
We should add some smarts about the order in which matches are
presented in the future, but for now having basic functionality
in place is still better than nothing ...