app: Make _ensureChannel() handle targetTypes other than ROOM
When unifying all callers to ensure_channel() with the 'join-room' action,
we broke the /query command which requires Tp.HandleType.CONTACT rather
than Tp.HandleType.ROOM.
Update _ensureChannel() to use an additional targetHandleType parameter
in its request data, so we can reuse it to implement a new 'message-user'
action (which we will also use eventually to implement the corresponding
menu action).

Based on a patch by Volker Sobek.