ui: Fix CSD window decorations
Tricky. GTK+ automatically sets the .titlebar style on custom titlebars,
which will give us a nice gradient background and rounded corners in
Adwaita. Except that GtkBox we are using as container for the two
HeaderBars does not have a window to paint on, so our titlebar ends
up with the default header-bar style - no gradient, no rounded corners.
Work around this by manually adding the .titlebar class to both
HeaderBars and use custom CSS to remove the unwanted border-radii.