app: Only save channels after connecting successfully
Connection failures may be either temporary (for instance when the
requested network is currently unavailable) or permanent (as in case
of configuration errors). Currently when saving channels, we sill still
try to remember the latter and try to (unsuccessfully) connect to them
each time we are started.
While we could try to disentangle the two classes of errors, only
saving channels on successful connection will take care of not ending
up with saved channels that are impossible to connect to, without
too bad drawbacks.