roomList: Explicitly select a new room when removing the current one
When the active room is removed, we currently update the selection
and rely on the row-selected handler to update the room; however we
disable this handler in selection mode (as selection has a very
different purpose there), so we end up with an inconsistent state
after removing the active room from selection mode. Fix this by
setting the new active room explicitly instead.