Change condition for enabling room list and selection mode switch
We currently disable those when the number of rooms drops to 0,
however as leaving a room can take a while, we hide rooms we are
leaving immediately to not feel sluggish - as a result, the room
list may remain empty for a while before it is hidden.
Adjust the condition to avoid this disconnect.