Disconnect accounts instead of leaving channels on shutdown
Leaving channels is horribly slow, resulting in a noticable disconnect
between closing the window and actually exiting the application.
Setting all active (IRC) accounts to offline instead is a lot faster,
so do that instead. While this is mostly a workaround for a telepathy-idle
issue, having an IRC connection running in the background is not actually
useful, so this arguably the correct behavior anyway.