chatView: Don't autoscroll to bottom while reading backlog
Scrolling automatically to the bottom as new messages arrive is
generally what we want, but it is a horrible disruption when it
kicks in while the user chose to scroll up to read up on previous
messages. To fix, only autoscroll while we are scrolled to the
bottom and keep the current scroll position otherwise.