mainWindow: Try harder to keep focus on the main entry
In general, we want the keyboard focus to remain on the main entry
so users don't have to explicitly focus it to write a message. The
only exceptions are:

 - focusing a different entry (nick entry, user list filter in the
   future)
 - using keynav

Unfortunately many widgets will grab the keyboard focus when activated,
and making them unfocusable would break accessibility. So resort to
a hack which forces focus back to the entry unless we detect one of
the exceptions above.

https://bugzilla.gnome.org/show_bug.cgi?id=709950