ui: Use a custom title widget in header bar
GtkHeaderBar's title/subtitle behavior is very close to what we want,
except that it does not allow markup for either property. However as
channel topics frequently contain URLs, this is something we depend on
in order to display URLs as clickable links, so replace GtkHeaderBar's
default title with a custom widget which replicates the built-in
behavior, but parses markup in subtitles.

https://bugzilla.gnome.org/show_bug.cgi?id=709889