roomList: Don't set 'dim-label' style on room icon
We used this style because the symbolic icons come out quite strong
with the default text color. Now that we dim inactive channels and
only use the normal color for channels with activity, this is less
of an issue - in fact, we should at least pick up the dimmed color
of inactive channels.
For now just let the icon pick up the text color, though we might
want to reconsider some dimming for active rooms.