accountsMonitor: Make primary owner of accountManager
Currently, both ChatroomManager and AccountsMonitor take a reference
on AccountsManager, with the former setting up the required features,
so it needs to be initialized first. Instead, handle AccountManager
initialization in AccountsMonitor and port ChatroomManager to use
it, so we can start to make use of AccountsMonitor's account tracking
in the future.

https://bugzilla.gnome.org/show_bug.cgi?id=732905