app: Automatically change to nominal nick when it becomes available
If a connection fails because the chosen nick is already in use, we
automatically retry with an appended underscore. Usually the reason
for this error is an unclean disconnect that leaves the nick from the
previous connection in zombie state until it times out - to get back
to the nominal nick, the user currently has to monitor the chat log
for the zombie timing out and manually change the nickname. We can
do better than that and use our user tracking to determine when the
nominal nick becomes available and change the nickname automatically
if there wasn't a user requested change in the meantime.

https://bugzilla.gnome.org/show_bug.cgi?id=710731