userList: Fix user count
We currently set/update the user count according to the number of
rows in the list. As we shortcut setting/updating the header in
case it's unneeded, we can end up with a count that is less than
the actual number of users.
To fix, base the counter on the number of users rather than widgets,
and make sure to always update it on room member changes.

https://bugzilla.gnome.org/show_bug.cgi?id=710655