chatView: Explicitly dispose cairo context
The context itself is managed by GTK+, but the JS wrapper object
is not. While the GC is expected to clean it up for us, it's buggy
and prone to leaks (which is fairly bad in a frequently called
function like draw()), so force a dispose() when no longer needed.