chatView: Slightly move mark for line-indicator
When marking a highlighted message, we want the mark at the end of
the newly inserted text, so we'll actually show the unread message
when scrolling it in view; the line indicator on the other hand
should appear above the entire message, which might be a different
display line when the message is wrapped. Set the mark on the start
of the line to get the right position.