chatView: Insert timestamps after inactivity
The time a message was sent/received is useful information, but
clutters the interface when repeated for every single message.
So rather than doing that (or chickening out by making it a user
option), only insert a timestamp after some time of inactivity,
which should allow to provide a good enough indication of when
a message was sent without the cluttering.

https://bugzilla.gnome.org/show_bug.cgi?id=712248