chatView: Use a separate tag for vertical gaps
We currently insert a vertical gap as part of the 'nick' and 'action'
tags. However we will start to insert timestamps for some messages,
in which case the gap should appear before the timestamp, not after
it. So create a dedicated tag for the vertical gap and insert it
where appropriate.

https://bugzilla.gnome.org/show_bug.cgi?id=712248