chatView: Fix condition for scrolling pending messages on screen
Currently the first pending message is scrolled on screen for
inactive rooms only, to avoid interfering with backlog reading.
This is a rather poor solution though, as auto-scrolling is the
desired behavior when not reading backlog. To fix, change the
condition to test for the current scroll position instead, as we
already do for non-highlighted messages.

https://bugzilla.gnome.org/show_bug.cgi?id=712217