room: Add account, channel-name and type properties
Currently the lifetime of a room is bound to the lifetime of the
corresponding channel. This will soon change, and we will create
a room as soon as the user requested it, and only destroy it when
it has been closed explicitly. To do that, we will need other
means to identify a room and associate it with the corresponding
channel, so add account, name and type properties for that purpose.
For now, those properties are read-only and derived from the channel,
but they will soon be used as the primary identifying properties.

https://bugzilla.gnome.org/show_bug.cgi?id=710271