mainWindow: Make entryArea a per-room widget
Currently when the main input entry contains text when switching
rooms, the text is taken over to the new room, which is unexpected.
Rather than keeping track of uncommitted input and saving/restoring
it on room switches, split out the entryArea code and use one for
each room as we already do for the chat log.