chatroomManager: Stop tying rooms' lifecycle to their channels
Currently rooms are created and destroyed as the corresponding
channel becomes available or goes away. However this behavior
can result in the application appearing sluggish when the time
between user request and result takes too long, and it is plain
wrong when no user interaction at all is involved (for instance
when the network goes away). So instead, create rooms immediately
when requested by the user, and keep them around until removed
explicitly.

https://bugzilla.gnome.org/show_bug.cgi?id=710271