app: Delete saved channels for deleted accounts
When an account is removed, its identifier may be reused for a
completely different connection. If that happens, it is rather
confusing to have "foreign" rooms (including history) restored.
There is nothing we can do when this happens behind our back,
but let's at least deal with it in case we can.

https://bugzilla.gnome.org/show_bug.cgi?id=715025