userList: Stop using a stack
By using a modal popover, the user list has now become a lot more
temporary. As a result, there is no longer a need for transitioning
between different lists on room switches, so we can just add the
correct list to the popover instead of keeping all lists around in
a stack. As GtkStack does a rather poor job at not propagating state
to non-visible children, this speeds up the UI significantly.

https://bugzilla.gnome.org/show_bug.cgi?id=727039