ChatroomManager: don't use two simple clients
What we want is one client that is both an observer and a handler.
We can't obtain that result with the simple classes, but we can
easily subclass TpBaseClient from gjs and implement both virtual functions.
This way, we only claim one name on the bus, and we only expose
one client path, which helps mission-control in routing messages
appropriately.

https://bugzilla.gnome.org/show_bug.cgi?id=723440