chatroomManager: Add 'active-state-changed' signal
In some places we care about the active room's status more than
about the room itself (e.g. whether the active room currently
has an associated channel or not). To track this, we'd have to
first listen for room changes and then track the room's channel
property - this is fairly cumbersome in most places, while trivial
from ChatroomManager where we are already tracking all rooms
anywhere, so add a new signal there.

https://bugzilla.gnome.org/show_bug.cgi?id=724902