chatroomManager: Remove rooms whose account has been disabled
Since rooms are no longer tied to their connection/channel, they are
only removed when the user explicitly requests it. This is clearly
the right behavior for "random" connection losses, but feels odd when
the room is disconnected because its account has been disabled - not
least as it is very much expected to be the result of user interaction.
Rooms with inactive accounts will also not be restored on startup,
so removing them when their account is disabled should also restore
consistency of the room list over restarts.