roomList: Move selection when the active room is removed
Since commit 0827a395dc24d22, rooms may be removed without the user
explicitly leaving them. If the currently active room is removed
like that, we will need to make sure to move the selection to not
end up with no active room and a hidden room list, just like we do
when explicitly leaving a room.