room: Base room icon on type rather than group interface
Private chat rooms use an icon to differentiate them from normal group
chats. Currently we look whether the channel implements the group
interface to determine whether to show the icon, but as channels
may come and go nowadays this doesn't work well anymore. Base the
icon on the (construct-only) type property instead to avoid
changing the icon based on connectivity.