chatView: Fix handling of channel changes
The current handling is sketchy at best - when the channel is not
unset entirely, we disconnect the old channel signal connections,
but using the already updated channel property. It does work because
a room's channel will be unset before it actually changes, but it's
still ugly. Just keep track of the channel separately, which also
allows to handle the case of the same channel being notified repeatedly
better than commit d71c13289adcaa14 did.