chatView: Also show notifications for highlights in non-active rooms
Currently we only show notifications while Polari is not focused.
However it makes sense to also show them when focused, when the
highlight was in a non-active room - while the room list will indicate
that there was a highlight, showing the actual message helps users
decide on whether to interrupt their current activity or not.