chatview: fix timestamp formatting
We want to show "yesterday" when the day part of the date is
different, even if less than 24 hours have passed. To do so,
instead of computing the difference between the actual time
points, use two reference points (located at 00:01) in the
same day, that thus will be always different by an integer
number of days.

https://bugzilla.gnome.org/show_bug.cgi?id=732611