ui: Select last used account by default in dialogs
Connections are currently ordered alphabetically with the first
one selected when opening the dialog. While not arbitrary, this
does not support users in joining the rooms they want as quickly
as possible - remembering the last account used for joining a room
and making it persistent has a higher chance of picking the account
the user wants, so do that instead of always selecting the first one.

https://bugzilla.gnome.org/show_bug.cgi?id=711833