userList: Don't keep userList widgets of inactive rooms
Since commit 9ab1ab5f09e3, we only keep the user list of the active
room in the widget hierarchy, which helped performance quite a bit.
However GTK+ still does a significant amount of work for unparented
widgets on style changes, so it is actually cheaper to destroy and
recreate the user list widget each time the active room changes.

Kudos to Matthias Clasen for the debugging legwork.

https://bugzilla.gnome.org/show_bug.cgi?id=732290