userList: Only create user list widget when necessary
We currently update the user list every time the active room changes.
However the user may never actually pop up the list, in which case we
are keeping a bunch of widgets around that are not shown at all - avoid
unnecessary work by only creating the list when we are about to show it.

https://bugzilla.gnome.org/show_bug.cgi?id=732290