userList: Only create user details on demand
Just like the user list as a whole, a user's details may never be
shown at all - in fact, for most users in the list this is the
expected case. As the user details (for each user in the list!)
contribute significantly to our excessive widget count, do the
same as for the user list and only create the details widget
when we are actually about to show it.

https://bugzilla.gnome.org/show_bug.cgi?id=732290