chatView: Update join time on channel changes
The code introduced in commit d2bbe17a76578e to prevent status
messages flooding the chat log when connecting broke when we
decoupled rooms and channels - rather than setting the join time
once on room creation, we need to update it every time we get
a connection established (i.e. a channel).

https://bugzilla.gnome.org/show_bug.cgi?id=736349