chatView: Block overshoot when fetching logs
GTK+ now displays an "overshoot indicator" when the users scrolls
past the top/bottom to indicate the end of the buffer. However
while we have history to load, scrolling past the top will prepend
to the buffer - indicate that we are not at the end in that case
by suppressing the overshoot indicator.

https://bugzilla.gnome.org/show_bug.cgi?id=740132