roomList: Show a context popover on right-click
Currently right-clicking a room will select the room and switch to
selection mode. However as it is very uncommon to leave more than
one room at once, so this behavior is far more cumbersome than
helpful. So instead, just open a context popover that allows to
directly leave the room in question.

https://bugzilla.gnome.org/show_bug.cgi?id=736481