Drop selection mode
In hindsight, the selection mode pattern was not a good choice for
room management - it only exposes a single action and is hardly used
for more than one item at once. In addition to this, the action uses
quite unfortunate wording - "Leave husband" - when applied to private
conversations (which are usually more volatile than rooms, thus more
commonly closed).
As the last commit made the leave action available without switching
to selection mode, just drop the pattern entirely now.

https://bugzilla.gnome.org/show_bug.cgi?id=736481