Add telepathy service and data files
We need to claim the Telepathy name as activatable in order
to be started by mission control when there is a channel
request to handle.

We claim we support all channels of type text, with all
handle types, like empathy-chat does. If we see a channel
of type != irc, we'll return an error and mission control
will transparently move to the next handler.

https://bugzilla.gnome.org/show_bug.cgi?id=745418