ChatView: update toplevel state when parent is set
state-flags-changed will be emitted only when the state flags
actually change, and therefore may not fire if the widget is
mapped to a window that is already focused.
In that case our tracking will get confused and think that
the window is not focused, resulting in messages that are
not properly acked.

https://bugzilla.gnome.org/show_bug.cgi?id=746508