roomOptions: Add room name, topic entry and leave room button
Adds a GtkLabel showing the display name of the current channel.
If the channel is a group channel, a separate box for group-chat
only options becomes visible. Currently the box only contains
an scrollable window with an editable textview with the group
chat topic. The current room can be left using the "Leave Room"
button in the bottom.