data: Add observer interface and filters to client file
When a telepathy client appears on the bus, mission-control will only use the
interface information exposed on the bus for the proxy object if it cannot
locate a .client file matching the client's bus name[0]. As we do not set
TpBaseClient:uniquify-name, our bus name does match the .client file added
in commit 21df75d20 and the interface information from the file will be used
instead.
However our .client file currently only announces the Handler interface, so
mission-control doesn't consider us an observer anymore - add appropriate
information for the Observer interface to fix that regression.

[0] http://cgit.freedesktop.org/telepathy/telepathy-mission-control/tree/src/mcd-client.c#n899

https://bugzilla.gnome.org/show_bug.cgi?id=752422