userList: Fix focusing the search entry after room changes
GtkContainer's default focus chain is based on its children's allocated
y position and height[0]. This is problematic when the user list has just
been created and doesn't have a proper allocation yet, as it will wrongly
sort before the search entry in that case.
Work around this by adding the user list to a dedicated bin (that is kept
around on room changes) instead of directly to the main box.

[0] https://git.gnome.org/browse/gtk+/tree/gtk/gtkcontainer.c#n2737

https://bugzilla.gnome.org/show_bug.cgi?id=750689