mainWindow: Fix min-width of user-list button
We currently set a width-request in the .ui file that used to be
consistent with the width of icon buttons, but the (default) theme
changed at one point, so this is no longer the case.
Fix this by setting the width-request to the actually allocated width
of an actual icon button (the join-menu).