roomList: Clear pending messages indication when offline
We currently update the indication for unread or highlighted messages
as the number of pending messages changes. However this obviously
does not work while offline, while at the same time keeping the
highlight around is obviously unobvious to users. The best thing
we can do while offline is clearing out the indication when a room
is selected, so do that.

https://bugzilla.gnome.org/show_bug.cgi?id=753050