chatroomManager: Handle authentication channels
Currently the only way to use servers that require a password with
polari is to set up the password elsewhere (Settings, Empathy, ...)
and rely on empathy's auth client to handle the actual authentication.
In order to support server passwords ourselves, we need to be able
to handle telepathy's password authentication mechanism before we offer
the user a way to specify the password.

https://bugzilla.gnome.org/show_bug.cgi?id=709824