room: Be more strict when highlighting messages
We currently highlight a messages if it contains the nick we match on,
regardless of where the match occurs. While this simplistic approach
works well enough for most nicks, it results in an annoying amount of
false positives for nicks that commonly appear in words. So tighten our
matching to not match when alphanumeric characters proceed or follow the
nick.

https://bugzilla.gnome.org/show_bug.cgi?id=755166