chatView: Base online/offline status on basenick
Since commit c96929c701bd, we use different colors for active and inactive
nicks. This has proven quite helpful when scrolling through the log, however
it can be confusing when multiple variants of a nick are present. As whether
a particular user is online at all is much more interesting than whether
they are currently identified by "foo" or "foo_", use a common status for
all nicks we think refer to the same person.

https://bugzilla.gnome.org/show_bug.cgi?id=755133