connection: Set a default response
Pressing enter in any of the entries in the entry should activate
the confirm button once the required fields have been filled out,
however that broke when the controls were moved into the dialog's
header bar in commit 96932bd60 as the button was no longer marked
as default. Set an appropriate default response to fix it again.