room: Improve channel-name <-> channel matching
When we request a channel, we currently assume that the resulting
channel's identifier will match the channel name we used in the
request. However as telepathy-idle converts all channel names to
lowercase [0], we fail to match names containing uppercase characters
to their channel and end up creating a second room for the channel
instead. Fix this by using lower-case for both room IDs and channel
matching.

[0]
http://cgit.freedesktop.org/telepathy/telepathy-idle/tree/src/idle-handles.c#n158

https://bugzilla.gnome.org/show_bug.cgi?id=755722