utils: Use paste.gnome.org as paste service
We only use the most basic functionality for uploading text to an
external paste service, so any paste service does the job for us
and fpaste is as good as any other from a technical point of view.
However politically it is problematic to use a service that is
provided by a specific distro, so pick a more neutral option with
the newly set up paste.gnome.org.

https://bugzilla.gnome.org/show_bug.cgi?id=726053