Add openURL helper to Utils
In order to bring the new application window to front
and get indication from the shell that we are opening an
application we need to use a GAppLaunchContext.

And it would be nice to get some feedback if we got
an error from opening the URL. For instance if no
handler is present.

https://bugzilla.gnome.org/show_bug.cgi?id=757493