tabCompletion: Fix completion popup on wayland
Unlike normal popups, the completion popup only serves as display for
completion options, so it does not grab input and should not take away
the focus from the corresponding entry. The latter doesn't currently
work on wayland, which makes us dismiss the popup immediately on map.

Work around this by marking the popup as tooltip (ugh), which allows
us to keep it working without taking a grab. In the future, we might
want to consider stop using a toplevel window for completions, for
instance a popover.

https://bugzilla.gnome.org/show_bug.cgi?id=758476