chatView: Fix updating compressed status message
When updating the status group header, we delete all text until
the end of the header tag before inserting the updated status
summary. However this means that on the first update, we remove
the newline between header and collapsed status message that is
included by the tag - fix by deleting to the line end instead of
the tag toggle.