chatView: Clean up handling of interactive tags
With the new status grouping where the header can be clicked to reveal
collapsed status messages, we now have a second text tag that responds
to input besides clickable links. This makes the code that handles them
centrally quite awkward, as we now need to differentiate not only on the
type of events, but also the tag type interacted with.
Clean this up by splitting out the generally useful event handling into
a ButtonTag class (with some help in TextView for hover-tracking), and
use its ::clicked, ::button-press-event and ::button-release-event signals
to implement links and status group headers and a much more natural way.

https://bugzilla.gnome.org/show_bug.cgi?id=759677