chatView: Animate auto-scrolling to bottom
Unless the user is reading the backlog or there are unread highlights,
we jump to the bottom of the view on arrival of new messages. While
the jump is usually a single line, animations generally give transitions
a more natural feel, so move the cursor instead which allows GTK+ to
animate the transition.

https://bugzilla.gnome.org/show_bug.cgi?id=759772