chatView: Improve link hover handling
Links are added with two text tags, one for generic styling shared
between all links and one url-specific one to handle interaction.
However with the hover styling added in commit cd69f6ce, we now
always override the foreground color specified by the generic tag.
This usually does not matter as the colors match, except that we
update the shared 'url' tag on style changes, but not the url-specific
ones. As hovering is a transient state, we don't care much about
updating hovered links immediately, but for links that are not
currently hovered, we shouldn't wait until the pointer moves over
the link again; simply stop overriding the 'url' tag in that case.