logManager: Add convenience wrapper for tracker queries
As telepathy-logger is unsuitable for anything more advanced than
retrieving logs sequentially (at least with acceptable performance),
we want to replace it with tracker eventually. To prepare for that,
add our own LogManager singleton that wraps a tracker connection and
provides a convenient API to get result objects from a SPARQL query.