logManager: Support running queries in chunks
It is not always desirable to retrieve the full result set at once;
in particular for the history in the chat log that uses TplLogWalker
currently, we want to maintain the behavior of only loading the most
recent history initially and fetch additional logs on demand.
To support this, extend the previously added GenericQuery class to
specify an optional limit.