app: Remove Connections dialog
After the latest changes, users can now create, modify and remove connections
without using the Connections dialog. Now the only functionality that is
exclusive to the dialog is the ability to disable connections temporarily,
which is only really useful during development - just move that to a dedicated
helper tool for that case and remove the dialog.