roomList: Extent headers to the full sidebar width
The current room header style is slightly odd: On the one hand, we use the
'row' CSS name to style it like a row (at this time: single background color
with no borders), on the other hand we add left and right margins that make
the header appear more like a button inside a row.
Resolve that inconsistency by removing the margins and slightly tweaking
the style to be more consistent with "normal" rows while maintaining the
current visuals when not active/hovered.

https://bugzilla.gnome.org/show_bug.cgi?id=761269