application: Allow to undo connection removal
Since commit 6ef2e57371aa added a "Remove" menu action to the room
header popover, it is now a lot easier to remove a connection by
accident. We should be more forgiving in that case, so show an app
notification that allows the user to undo the operation, and only
actually remove the account when the notification has been dismissed
or timed out.

https://bugzilla.gnome.org/show_bug.cgi?id=761281