roomList: Properly translate between room- and row index
Adding placeholders broke the assumption that the nth room in the
list corresponds to the nth row. Commit 36ed1c992 added some ad-hoc
handling to _moveSelection(), but we make the assumption in other
places as well - fix by adding helper functions to properly translate
between room- and row index and use those when appropriate.