entryArea: Don't grab focus unconditionally when becoming sensitive
Our current idea for supporting server passwords is to allow users to
enter a password on authentication errors using the header popovers in
the room list. As this may happen while the active room is connecting,
we should make sure to not steal the focus while the user is typing
her password.

https://bugzilla.gnome.org/show_bug.cgi?id=709824