joinDialog: Overhaul connection page
Use the newly added ConnectionsList to allow users to create a new
connection with a single click on a predefined network. For cases
where the server is not included in the list or more control is
needed, the previous UI for creating a connection manually can
still be accessed through a "Custom Connection" toggle.

https://bugzilla.gnome.org/show_bug.cgi?id=761859