connections: Move error information into properties dialog
The ConnectionDetails widget is used both for creating new connections
in the join dialog and for editing existing ones from the properties
dialog, but only in the latter case we may show the error information.
However it still imposes spacing on the details grid even when hidden,
which makes it harder to adapt the grid to the different parents in the
future - just move the error information into the dialog that uses it.

https://bugzilla.gnome.org/show_bug.cgi?id=761859