connections: Hide server settings for predefined networks
The whole point of providing a list of networks is so users needn't
bother with server addresses and ports. In addition, exposing those
settings blurs the line between predefined and custom connections,
which is problematic for features like using encrypted connections
where possible.

So just hide server settings for accounts created from a predefined
network - in the end, users who are unhappy with the picks we make
can still use a custom connection.

https://bugzilla.gnome.org/show_bug.cgi?id=761859