Add NetworksManager to manage a list of predefined networks
Picking a predefined entry instead of having to enter a hostname/IP and
possibly a port will be a big usability improvement for users. And for us,
knowing which network the users wants to connect to will open the door for
future improvements, like using encrypted connections where available or
having a list of fallback servers to try on connection failures.

The list itself is mostly irc-networks.xml from telepathy-account-widgets
converted to JSON (because gjs' XML story sucks) - there's a fair chance
that entries are as outdated as the GNOME ones were, so take it with a
grain of salt ...

https://bugzilla.gnome.org/show_bug.cgi?id=761859