joinDialog: Allow filtering predefined connections
The list of connections is quite extensive, which makes it hard to
navigate. Address this by adding a filter entry which follows the
preferred search pattern: The searched text is split into words and
only items that match all terms are displayed, which allows to
quickly narrow down the number of items.

https://bugzilla.gnome.org/show_bug.cgi?id=761859