roomList: Fix leaking room headers
When unsetting a row's header by passing null to set_header(), the row will
drop the reference it holds on the existing header, but not destroy it.
As a consequence, the reference we hold ourselves because of connected
signal handlers will keep the object alive indefinitely - make sure the
header is properly destroyed to fix this leak.