roomStack: Add placeholder directly
Since commit 3f9eea26f, we no longer show an entry area with the placeholder.
This was done in RoomView by replacing the "proper" entry area with an empty
container - however the size group will give that container a height once a
room (and thus a real entry area) is added, with the result of the chat
placeholder jumping noticeably during the transition.
To fix, overlay the pseudo entry area over the placeholder and add the latter
directly to the stack instead of going through RoomView.