entryArea: Use pasteManager directly to handle pastes
Our current paste handling works pretty much in a fire-and-forget way:
the pasted content is handed over to the pasteManager which will take
care of uploading it and sending the resulting URL to the current channel.
As a result, URLs are always inserted as standalone message with no way
for users to add additional context like highlights.
To address this, the entry will need access to the returned URL before
the message is send - achieving this using actions that don't have
a return value would be messy, so simply cut out the middle man and
use the pasteManager singleton directly.

https://bugzilla.gnome.org/show_bug.cgi?id=760315