entryArea: Insert paste URL instead of sending it directly
It usually makes sense to provide some context when sending a paste
URL, so allow that by inserting the returned paste URL into the
entry instead of sending it directly. Pasting content as URL is
now a two-step operation - confirming the upload and sending the
message - but hopefully that'll just be minor annoyance compared
to the usefulness of the feature.

https://bugzilla.gnome.org/show_bug.cgi?id=760315