entryArea: Use paste-confirmation area to indicate upload progress
Now that we insert paste URL into the entry rather than sending them
directly, we don't want the current entry text to change while the
upload is in progress - we'd just end up with situations where the
insertion interferes with what the user is currently typing.

To prevent this, we only switch back to the entry after the upload
has finished, which makes the paste-confirmation area available for
replacing the old upload notification we removed earlier.

https://bugzilla.gnome.org/show_bug.cgi?id=760315