room: Consider account nick for highlighting
Using the basenick to group nicks that refer to the same identity works
fairly well for most nicks. A notable exceptions are multi-word nicks
that use dashes or underscores as separator. We can't do much in that
case for other contacts where all we have is the nickname, but in case
of the user's own nick we have more data points we can use to get a
better result - namely the nickname that has been configured for the
account.

Use that to improve the accuracy of message highlighting, which
constitutes the main annoyance for affected users.

https://gitlab.gnome.org/GNOME/polari/issues/69