connectionPopover: Reconnect does not do anything when connected
When already connected, the reconnect button on popover does not do
anything.

To fix this issue, we changed the reconnect action which is now
using  the  reconnect_async() from tp account to disconnect and
connect in every call. But this solution doesn't apply when the
account is offline, the  proposed solution for this in a future
implementation is to connect to all accounts on startup.

https://bugzilla.gnome.org/show_bug.cgi?id=761276